+++ src/app.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Main from './pages/Main/main';
-import AdminPanel from './pages/AdminPanel/admin';
+import Main from './pages/main/main';
+import AdminPanel from './pages/admin-panel/admin';
 import Card from './pages/worker-card';
 import Edit from './pages/card-editing';
