+++ src/components/Card/index.ts
@@ -1,1 +1,1 @@
-export { Card } from './Card';
+export { Card } from './card';
