+++ src/components/Card/styled.ts
