+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import { Burger } from "./Burger";
-import { HeaderLinks } from "../HeaderLinks";
+import { Burger } from "./burger";
+import { HeaderLinks } from "../header-links";
 
 import logo from '../../assets/icons/logo.svg';
 
@@ -15,7 +15,7 @@ export function Header() {
     <header className="header">
       <div className="header__container">
         <a className="header__logo logo" href="#">
-          <img className="logo__icon" src={logo} alt="WebStar" />
+          <img className="logo__icon" src={logo} srcSet={logo} alt="WebStar" />
         </a>
         <HeaderLinks isOpen={isHeaderLinksOpen} />
         <Burger onHeaderLinksToggle={handleHeaderLinksToggle} />
