+++ src/components/Header/index.ts
@@ -1,1 +1,1 @@
-export { Header } from "./Header";
+export { Header } from "./header";
