--- src/components/HeaderLinks/index.ts
@@ -1,1 +1,0 @@
-export { HeaderLinks } from './HeaderLinks';
