+++ src/components/Message/Message.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { MessageButtons } from "./MessageButtons";
+import { MessageButtons } from "./message-buttons";
 import { MessageDiv, MessageNum, Title } from "./styled";
 
 export function Message({ number, title }) {
