+++ src/components/Message/index.ts
@@ -1,1 +1,1 @@
-export { Message } from './Message';
+export { Message } from './message';
