+++ src/components/Search/index.ts
@@ -1,1 +1,1 @@
-export { Search } from './Search';
+export { Search } from './search';
