+++ src/components/Search/styled.ts
