+++ src/components/Sidebar/index.ts
@@ -1,1 +1,1 @@
-export { Sidebar } from './Sidebar'
+export { Sidebar } from './sidebar'
