+++ src/components/header-links/index.ts
@@ -1,0 +1,1 @@
+export { HeaderLinks } from './header-links';
