+++ src/components/title/title.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { MainTitle } from "./styled";
+
+export function Title(props) {
+  return <MainTitle letterSpace={props.letterSpace}>{props.children}</MainTitle>
+}
