--- src/script.js
@@ -1,33 +1,0 @@
-// const burgerBgContent = document.querySelector('.burger-bg__content');
-// const burger = document.getElementById("burger");
-// const headerLinks = document.querySelector(".header__links");
-// const burgerBg = document.getElementById("burger-bg");
-// const body = document.body;
-
-// function burgerOnClick() {
-//   burgerBgContent.classList.toggle("change");
-//   burger.classList.toggle("change");
-//   headerLinks.classList.toggle("change");
-//   burgerBg.classList.toggle("change-bg");
-// }
-
-// body.addEventListener("click", function(e) {
-//   if (burger.classList.contains('change') && e.target !== burger && !burger.contains(e.target)) {
-//     burgerOnClick();
-//   }
-// });
-
-// const sidebarBtn = document.querySelector('.sidebar__btn');
-// const sidebar = document.querySelector('.sidebar');
-// sidebarBtn.addEventListener("click", () => {
-//   sidebar.classList.toggle("_open");
-//   sidebarBtn.classList.toggle("_open");
-// })
-
-// if (window.innerWidth >= 1024) {
-//   setTimeout(() => {
-//     sidebar.classList.add("_open");
-//     sidebarBtn.classList.add("_open");
-//   } ,400);
-  
-// }
